% Auto-generated: do not edit by hand
\name{componentB}

\alias{componentB}

\title{ComponentB component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
componentB(children=NULL, id=NULL, a=NULL, b=NULL, c=NULL, d=NULL,
e=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Children}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{a}{Character. A}

\item{b}{Character. A}

\item{c}{Character. A}

\item{d}{Character. A}

\item{e}{Character. A}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

